+++ CHANGE_LOG.md
@@ -6,6 +6,10 @@
 - **CardContent**: Create the card content component.
 - **CardFooter**: Create the card footer component.
 - **CardHeader**: Create the card header component.
+- **RadionButton**: added NEW `RadioButton` component
+- **RadioGroup**: added NEW `RadioGroup` component
+- **RadionButton**: moved to Components
+- **RadioGroup**: moved to Components
 
 ## 0.0.0-alpha.8